# internal/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Logiflow API
  description: API для логистической информационной системы
  version: 0.2.0
  contact:
    name: anxi0uz
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          nullable: false
        data:
          type: object
          nullable: true
        success:
          type: boolean
          nullable: false
        requestID:
          type: string
          nullable: false
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        requestID:
          type: string

    # ─── Auth ─────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 150
          nullable: true
        avatarUrl:
          type: string
          format: uri
          nullable: true
        password:
          type: string
          minLength: 8
          nullable: true
        currentPassword:
          type: string
          minLength: 8
          nullable: true
    UserDeleteRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string

    # ─── Vehicle ──────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      summary: Регистрация клиента
      description: Публичный эндпоинт — создаёт аккаунт с ролью client. Для создания менеджеров и водителей используйте POST /managers и POST /drivers (только admin).
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Пользователь создан, токены выданы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email уже занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: authLogin
      summary: Авторизация
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: authRefresh
      summary: Обновление access-токена
      tags: [Auth]
      security: []
      responses:
        "200":
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Недействительный refresh-токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: authLogout
      summary: Выход из системы
      tags: [Auth]
      responses:
        "200":
          description: Успешный выход
        "401":
          description: Не авторизован

  # ─── Me ─────────────────────────────────────────────────────────────────────
  /me:
    get:
      operationId: getMe
      summary: Текущий пользователь
      tags: [Me]
      responses:
        "200":
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Не авторизован

    patch:
      operationId: updateMe
      summary: Обновить профиль
      tags: [Me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Профиль обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Не авторизован

    delete:
      operationId: deleteMe
      summary: Удалить аккаунт
      tags: [Me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDeleteRequest"
      responses:
        "200":
          description: Аккаунт удалён
        "400":
          description: Неверный пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Не авторизован
  /me/trips:
    get:
      operationId: getMyTrips
      summary: Поездки текущего водителя
      tags: [Me]
      description: Доступно только для роли driver. Возвращает маршруты, назначенные на текущего водителя.
      responses:
        "200":
          description: Список маршрутов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Доступно только водителям

  # ─── Vehicles ───────────────────────────────────────────────────────────────
...